--- backend/commands/seed.go
@@ -1,9 +1,0 @@
-package commands
-
-// func Seed(c *fiber.Ctx) error {
-// 	err := database.SeedDB(client)
-// 	if err != nil {
-// 			return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"message": "Error seeding database", "error": err.Error()})
-// 	}
-// 	return c.JSON(fiber.Map{"message": "Seeding done!"})
-// }
